+++ src/plugins/yagr/renderer/tooltip/index.ts
@@ -1,0 +1,1 @@
+export {renderTooltip} from './renderTooltip';
